Add Tall colors and fix score text.

+++ Assets/Scripts/GameManager.cs
@@ -68,7 +68,7 @@ public class GameManager : MonoBehaviour
         
         endMenu.gameObject.SetActive(true);
         endMenu.ShowResult(expressionIndex);
-        endMenuScore.SetText("Dorothy's happiness: " + scoreResolver.Score);
+        endMenuScore.SetText("Dorothy's happiness: " + (int)scoreResolver.Score);
         endMenuDistance.SetText("You made " + scoreResolver.TotalDistance.ToString("0.00") + "m of dominos!");
         
         audioSource.clip = audioClips[expressionIndex];
